simple-flask-blog/pyproject.toml
[project]
name = "simple-flask-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.3.0",
    "python-dotenv>=1.2.2",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]  # error, flake, import

[tool.ruff.format]
quote-style = "double"
line-ending = "cr-lf"

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
